Keep tests in sync with HEAD.

+++ lustre/tests/sanity.sh
@@ -1240,14 +1240,19 @@ stop_kupdated() {
 }
 
 # ensure that all stripes have some grant before we test client-side cache
-for i in `seq -f $DIR/f42-%g 1 $STRIPECOUNT`; do
-	dd if=/dev/zero of=$i bs=4k count=1
-	rm $i
-done
+setup_test42() {
+	[ "$SETUP_TEST42" ] && return
+	for i in `seq -f $DIR/f42-%g 1 $STRIPECOUNT`; do
+		dd if=/dev/zero of=$i bs=4k count=1
+		rm $i
+	done
+	SETUP_TEST42=DONE
+}
 
 # Tests 42* verify that our behaviour is correct WRT caching, file closure,
 # file truncation, and file removal.
 test_42a() {
+	setup_test42
 	cancel_lru_locks OSC
 	stop_kupdated
 	sync; sleep 1; sync # just to be safe
@@ -1262,6 +1267,7 @@ test_42a() {
 run_test 42a "ensure that we don't flush on close =============="
 
 test_42b() {
+	setup_test42
 	cancel_lru_locks OSC
 	stop_kupdated
         sync
